requirements.txt
crewai>=0.76.2
litellm>=1.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
langchain>=0.1.0